Manage ignition server from CPO
This changes moves the ignition server management from the HO into the
CPO. In order to stay backwards compatible, the HO will still reconcile
the ignition server if the CPO isn't opted into doing that through a
Docker image label.

Ref https://issues.redhat.com/browse/HOSTEDCP-411
