+++ control-plane-operator/controllers/hostedcontrolplane/ignitionserver/ignitionserver_test.go
@@ -1,0 +1,142 @@
+package ignitionserver
+
+import (
+	"testing"
+
+	. "github.com/onsi/gomega"
+	hyperv1 "github.com/openshift/hypershift/api/v1alpha1"
+	"github.com/openshift/hypershift/hypershift-operator/controllers/manifests/ignitionserver"
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/util/intstr"
+)
+
+func TestReconcileIgnitionServerServiceNodePortFreshInitialization(t *testing.T) {
+	tests := []struct {
+		name                           string
+		inputIgnitionServerService     *corev1.Service
+		inputServicePublishingStrategy *hyperv1.ServicePublishingStrategy
+	}{
+		{
+			name:                       "fresh service initialization",
+			inputIgnitionServerService: ignitionserver.Service("default"),
+			inputServicePublishingStrategy: &hyperv1.ServicePublishingStrategy{
+				Type: hyperv1.NodePort,
+			},
+		},
+		{
+			name:                       "fresh service with node port specified",
+			inputIgnitionServerService: ignitionserver.Service("default"),
+			inputServicePublishingStrategy: &hyperv1.ServicePublishingStrategy{
+				Type: hyperv1.NodePort,
+				NodePort: &hyperv1.NodePortPublishingStrategy{
+					Port: int32(30000),
+				},
+			},
+		},
+	}
+	for _, test := range tests {
+		t.Run(test.name, func(t *testing.T) {
+			reconcileIgnitionServerService(test.inputIgnitionServerService, test.inputServicePublishingStrategy)
+			g := NewGomegaWithT(t)
+			g.Expect(len(test.inputIgnitionServerService.Spec.Ports)).To(Equal(1))
+			g.Expect(test.inputIgnitionServerService.Spec.Type).To(Equal(corev1.ServiceTypeNodePort))
+			g.Expect(test.inputIgnitionServerService.Spec.Ports[0].TargetPort).To(Equal(intstr.FromInt(9090)))
+			g.Expect(test.inputIgnitionServerService.Spec.Ports[0].Port).To(Equal(int32(443)))
+			g.Expect(test.inputIgnitionServerService.Spec.Ports[0].Name).To(Equal("https"))
+			g.Expect(test.inputIgnitionServerService.Spec.Ports[0].Protocol).To(Equal(corev1.ProtocolTCP))
+			if test.inputServicePublishingStrategy.NodePort != nil && test.inputServicePublishingStrategy.NodePort.Port > 0 {
+				g.Expect(test.inputIgnitionServerService.Spec.Ports[0].NodePort).To(Equal(test.inputServicePublishingStrategy.NodePort.Port))
+			}
+		})
+	}
+}
+
+func TestReconcileIgnitionServerServiceNodePortExistingService(t *testing.T) {
+	tests := []struct {
+		name                           string
+		inputIgnitionServerService     *corev1.Service
+		inputServicePublishingStrategy *hyperv1.ServicePublishingStrategy
+	}{
+		{
+			name: "existing service keeps nodeport",
+			inputIgnitionServerService: &corev1.Service{
+				ObjectMeta: ignitionserver.Service("default").ObjectMeta,
+				Spec: corev1.ServiceSpec{
+					Ports: []corev1.ServicePort{
+						{
+							Name:       "https",
+							Port:       443,
+							TargetPort: intstr.FromInt(9090),
+							Protocol:   corev1.ProtocolTCP,
+							NodePort:   int32(30000),
+						},
+					},
+				},
+			},
+			inputServicePublishingStrategy: &hyperv1.ServicePublishingStrategy{
+				Type: hyperv1.NodePort,
+			},
+		},
+	}
+	for _, test := range tests {
+		t.Run(test.name, func(t *testing.T) {
+			initialNodePort := test.inputIgnitionServerService.Spec.Ports[0].NodePort
+			reconcileIgnitionServerService(test.inputIgnitionServerService, test.inputServicePublishingStrategy)
+			g := NewGomegaWithT(t)
+			g.Expect(len(test.inputIgnitionServerService.Spec.Ports)).To(Equal(1))
+			g.Expect(test.inputIgnitionServerService.Spec.Type).To(Equal(corev1.ServiceTypeNodePort))
+			g.Expect(test.inputIgnitionServerService.Spec.Ports[0].TargetPort).To(Equal(intstr.FromInt(9090)))
+			g.Expect(test.inputIgnitionServerService.Spec.Ports[0].Port).To(Equal(int32(443)))
+			g.Expect(test.inputIgnitionServerService.Spec.Ports[0].Name).To(Equal("https"))
+			g.Expect(test.inputIgnitionServerService.Spec.Ports[0].Protocol).To(Equal(corev1.ProtocolTCP))
+			g.Expect(test.inputIgnitionServerService.Spec.Ports[0].NodePort).To(Equal(initialNodePort))
+		})
+	}
+}
+
+func TestReconcileIgnitionServerServiceRoute(t *testing.T) {
+	tests := []struct {
+		name                           string
+		inputIgnitionServerService     *corev1.Service
+		inputServicePublishingStrategy *hyperv1.ServicePublishingStrategy
+	}{
+		{
+			name:                       "fresh service initialization",
+			inputIgnitionServerService: ignitionserver.Service("default"),
+			inputServicePublishingStrategy: &hyperv1.ServicePublishingStrategy{
+				Type: hyperv1.Route,
+			},
+		},
+		{
+			name: "existing service",
+			inputIgnitionServerService: &corev1.Service{
+				ObjectMeta: ignitionserver.Service("default").ObjectMeta,
+				Spec: corev1.ServiceSpec{
+					Ports: []corev1.ServicePort{
+						{
+							Name:       "https",
+							Port:       443,
+							TargetPort: intstr.FromInt(9090),
+							Protocol:   corev1.ProtocolTCP,
+						},
+					},
+				},
+			},
+			inputServicePublishingStrategy: &hyperv1.ServicePublishingStrategy{
+				Type: hyperv1.Route,
+			},
+		},
+	}
+	for _, test := range tests {
+		t.Run(test.name, func(t *testing.T) {
+			reconcileIgnitionServerService(test.inputIgnitionServerService, test.inputServicePublishingStrategy)
+			g := NewGomegaWithT(t)
+			g.Expect(len(test.inputIgnitionServerService.Spec.Ports)).To(Equal(1))
+			g.Expect(test.inputIgnitionServerService.Spec.Type).To(Equal(corev1.ServiceTypeClusterIP))
+			g.Expect(test.inputIgnitionServerService.Spec.Ports[0].TargetPort).To(Equal(intstr.FromInt(9090)))
+			g.Expect(test.inputIgnitionServerService.Spec.Ports[0].Port).To(Equal(int32(443)))
+			g.Expect(test.inputIgnitionServerService.Spec.Ports[0].Name).To(Equal("https"))
+			g.Expect(test.inputIgnitionServerService.Spec.Ports[0].Protocol).To(Equal(corev1.ProtocolTCP))
+		})
+	}
+}
